+++ src/Yaat.Sim/Commands/ParsedCommand.cs
@@ -85,7 +85,8 @@ public record ForceSpeedCommand(int Speed) : ParsedCommand;
 public record WarpCommand(string PositionLabel, double Latitude, double Longitude, MagneticHeading? MagneticHeading, int? Altitude, int? Speed)
     : ParsedCommand;
 
-public record WarpGroundCommand(string Taxiway1, string Taxiway2, int? NodeId = null, string? ParkingName = null) : ParsedCommand;
+public record WarpGroundCommand(string Taxiway1, string Taxiway2, int? NodeId = null, string? ParkingName = null, string? SpotName = null)
+    : ParsedCommand;
 
 public record SquawkCommand(uint Code) : ParsedCommand;